--- lib/foods.js
@@ -1,4 +1,0 @@
-//create variable to call api and store return
-//parse return
-//invoke variable on foods.html
-//iterate through collection and populate table
